backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ─────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# ── Cache / Queue ─────────────────────────────────────────────
redis[hiredis]==5.0.4
celery[redis]==5.4.0

# ── Auth & Security ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# ── HTTP & Utils ─────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1
Pillow==10.3.0
boto3==1.34.102

# ── Notifications ─────────────────────────────────────────────
twilio==9.0.4
sendgrid==6.11.0

# ── AI / ML ───────────────────────────────────────────────────
openai==1.30.1
anthropic==0.26.0
sentence-transformers==3.0.1

# ── Scraping ──────────────────────────────────────────────────
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
scrapy==2.11.2

# ── Testing ───────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0

# ── Payments ──────────────────────────────────────────────────────────────────
razorpay==1.4.2
